pyproject.toml
[project]
name = "miia-question-validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.48.0",
    "google-genai>=1.64.0",
    "gspread>=6.2.1",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
